jobs/isetan-data-scraping/requirements.txt
joblib>=1.5.2
jpholiday>=1.0.2
matplotlib>=3.10.6
numpy<2
openpyxl>=3.1.5
pandas>=2.3.2
plotly>=6.3.0
pyyaml>=6.0.3
requests>=2.32.5
scikit-learn>=1.7.2
seaborn>=0.13.2
shap>=0.48.0
xgboost>=3.0.5
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-api-python-client>=2.146.0
google-auth>=2.35.0
google-cloud-storage>=2.18.0
google-cloud-secret-manager>=2.21.0
google-cloud-bigquery>=3.0.0
pandas-gbq>=0.26.1
pyarrow>=16.0.0
playwright>=1.40.0
pytz>=2025.2
